Refactor: Move inclusion and declaration to header
