+++ src/reference.cpp
@@ -1,11 +1,6 @@
-#include <iostream>
+#include "reference.h"
 
 using namespace std;
-
-template<typename T>
-T sum(T a, T b);
-
-void swap(int& a, int& b);
 
 int main() {
     int a = 100;
@@ -29,10 +24,6 @@ int main() {
     cout << "call by reference can successfully swap two numbers";
 }
 
-template<typename T>
-T sum(T a, T b) {
-    return a + b;
-}
 
 void swap(int& a, int& b) {
     int temp = a;
